Break out dependency on getActivePlatform
