+++ src/app/destinyTrackerApi/d2-reviewsFetcher.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { getActivePlatform } from '../accounts/platforms';
+import { getActivePlatform } from '../accounts/get-active-platform';
 import { loadingTracker } from '../shell/loading-tracker';
 import { handleD2Errors } from './d2-trackerErrorHandler';
 import { D2Item } from '../inventory/item-types';
