+++ src/app/settings/SettingsPage.tsx
@@ -18,7 +18,8 @@ import { D2StoresService } from '../inventory/d2-stores';
 import { D1StoresService } from '../inventory/d1-stores';
 import Checkbox from './Checkbox';
 import Select, { mapToOptions, listToOptions } from './Select';
-import { getPlatforms, getActivePlatform } from '../accounts/platforms';
+import { getPlatforms } from '../accounts/platforms';
+import { getActivePlatform } from '../accounts/get-active-platform';
 import { itemSortOrder } from './item-sort';
 import { Settings } from './initial-settings';
 import { settingsSelector } from './reducer';
